[Util-Calendar] Separate Ui Components
