+++ util/calendar/src/main/java/com/example/calendar/model/CalendarDate.kt
@@ -1,0 +1,8 @@
+package com.example.calendar.model
+
+data class CalendarDate(
+    val year: Int,
+    val month: Int,
+    val day: Int,
+    val dayOfWeek: String
+)
